Converter: route named-numeric↔basic/named conversions through the underlying
A Go conversion between a named numeric type and a DIFFERENT basic numeric
(`uint64(nameOff)`, `int(idx)`) or another named numeric (`hex(nameOff)`) emitted
a plain `(ulong)NameOff` / `(Δhex)(uint64)NameOff` cast. The [GoType] wrapper only
converts between the named type and its EXACT underlying, so those are CS0030.

convCallExpr now routes through the underlying, mirroring the existing "convert TO
a named numeric" path:
- FROM a named numeric to a different basic: `((uint64)(int32)off)` (new branch).
- TO a named numeric whose ARG is itself a named numeric with a different
  underlying: unwrap the arg first — `((Δhex)(uint64)(int32)off)`.

Key gotcha: runtime's `nameOff`/`typeOff`/`textOff` are Go type ALIASES
(`type nameOff = abi.NameOff`), so `info.TypeOf(arg)` returns `*types.Alias` in
Go 1.23 — `types.Unalias` before the `*types.Named` assertion.

Runtime 197→187 (CS0030 58→48); the type.cs abi-offset, tracecpu, tagptr_64bit
sites cleared. Full behavioral suite green (186). Test: extended NamedNumericConversion
(named→basic both directions + named→named through both underlyings). Intended
through-underlying churn re-baselined in IotaEnum + StdLibInternalAbi (`int(ΔKind)`
→ `(nint)(uint8)k`). Doc: ConversionStrategies "Converting from a named numeric".

+++ src/Tests/Behavioral/StdLibInternalAbi/type.cs
@@ -67,7 +67,7 @@ public static readonly TFlag TFlagUnrolledBitmap = /* 1 << 4 */ 16;
 [GoType("num:int32")] partial struct TextOff;
 
 public static @string String(this ΔKind k) {
-    if (((nint)k) < len(kindNames)) {
+    if (((nint)(uint8)k) < len(kindNames)) {
         return kindNames[k];
     }
     return kindNames[0];
